requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
async-timeout==4.0.2
autopep8==1.5.7
beautifulsoup4==4.11.1
boto3==1.21.24
botocore==1.24.24
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
confuse==1.7.0
cryptography==36.0.0
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-allauth==0.49.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-localflavor==3.1
django-multiselectfield==0.1.12
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
fonttools==4.32.0
geographiclib==1.52
geopy==2.2.0
google-ads==2.0.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.53.0
grpcio==1.44.0
gunicorn==20.1.0
httpie==2.6.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.8.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==1.0.0
jwcrypto==1.0
jws==0.1.3
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.2
mysqlclient==2.1.0
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.7
packaging==20.9
pandas==1.4.2
Pillow==9.0.1
protobuf==3.15.7
psycopg2==2.9.3
psycopg2-binary==2.9.2
py-moneyed==0.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-amazon-sp-api==0.13.7
python-dateutil==2.8.2
python-decouple==3.5
python-jwt==2.0.1
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzipcode==3.0.1
redis==4.2.2
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.2
selenium==3.141.0
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.2
SQLAlchemy==1.3.17
sqlparse==0.4.1
sseclient==0.0.27
suds-jurko==0.6
toml==0.10.2
tqdm==4.61.2
unicodecsv==0.13.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.2
whitenoise==5.3.0
WooCommerce==3.0.0
wrapt==1.14.0
WTForms==2.3.3
xlwt==1.3.0
zipp==3.6.0
